requirements.txt
pytest
pytest-html
numpy
pandas
scikit-learn
pyyaml